+++ backend/src/auth/auth.controller.ts
@@ -1,4 +1,4 @@
-import { Body, Controller, Get, Post, Headers, UseGuards } from "@nestjs/common";
+import { Body, Controller, Get, Post, Headers, UseGuards, Param } from "@nestjs/common";
 import { CreateUserDto } from "../user/dto/create-user.dto";
 import { AuthService } from "./auth.service";
 import { LoginUserDto } from "../user/dto/login-user.dto";
@@ -6,6 +6,7 @@ import { ApiOperation, ApiTags, ApiOkResponse } from "@nestjs/swagger";
 import { UserModel } from "../user/models/user.model";
 import { RolesGuard } from "./roles.guard";
 import { Roles } from "./role-auth.decorator";
+import { ForgotPasswordDto } from "../user/dto/forgot-password.dto";
 
 @ApiTags("Auth functional")
 @Controller("auth")
@@ -39,5 +40,14 @@ export class AuthController {
   @Get("/me")
   me(@Headers() headers) {
     return this.authService.me(headers.authorization);
+  }
+
+  @ApiOperation({ summary: "Is current user authorized" })
+  @ApiOkResponse({ status: 200, type: UserModel })
+  @Roles("ADMIN", "USER")
+  @UseGuards(RolesGuard)
+  @Get("/forgotPassword")
+  forgotPassword(@Body() dto: ForgotPasswordDto) {
+    return this.authService.forgotPassword(dto.email);
   }
 }